Moved call to /profile endpoint to LoggedInLayout.js
